+++ Mailer/app/Forms/MediaBriefingTemplateFormFactory.php
@@ -108,31 +108,12 @@ class MediaBriefingTemplateFormFactory
         return $form;
     }
 
-    private function getUniqueTemplateCode(string $code): string
-    {
-        $storedCodes = $this->templatesRepository->getTable()
-            ->where('code LIKE ?', $code . '%')
-            ->select('code')->fetchAll();
-
-        if ($storedCodes) {
-            $max = 0;
-            foreach ($storedCodes as $c) {
-                $parts = explode('-', $c->code);
-                if (count($parts) > 1) {
-                    $max = max($max, (int) $parts[1]);
-                }
-            }
-            $code .= '-' . ($max + 1);
-        }
-        return $code;
-    }
-
     public function formSucceeded(Form $form, ArrayHash $values): void
     {
         $generate = function ($htmlBody, $textBody, $mailLayoutId, $segmentCode = null) use ($values, $form) {
             $mailTemplate = $this->templatesRepository->add(
                 $values['name'],
-                $this->getUniqueTemplateCode($values['code']),
+                $this->templatesRepository->getUniqueTemplateCode($values['code']),
                 '',
                 $values['from'],
                 $values['subject'],